Disable of adding proxy settings for apt
At this moment we have two tools which do same thing but different ways.
There is setup proxy address for apt cache. The first one is debootstrap
which adding proxy ip addresses to sources.list and the second one is
this cookbook which setup proxy address in sources.list.d folder.
Now we remove this part from recipe for correct work of apt.